Move thread rendering from `ThreadComponent` to `ThreadView`
